Added ConsoleRenderer and tests
