+++ GameOfLifeTests/GameOfLifeTests.csproj
@@ -9,7 +9,7 @@
     <CollectCoverage>true</CollectCoverage>
     <CoverletOutputFormat>cobertura</CoverletOutputFormat>
     <CoverletOutput>TestResults/coverage.cobertura.xml</CoverletOutput>
-    <Threshold>95</Threshold>
+    <Threshold>85</Threshold>
     <ThresholdType>line</ThresholdType>
     <ThresholdStat>total</ThresholdStat>
   </PropertyGroup>
